Move expression suggest files out of metabase-lib
