+++ frontend/src/metabase/query_builder/components/expressions/ExpressionEditorHelpText.tsx
@@ -6,8 +6,8 @@ import MetabaseSettings from "metabase/lib/settings";
 import ExternalLink from "metabase/core/components/ExternalLink";
 import Icon from "metabase/components/Icon";
 import TippyPopover from "metabase/components/Popover/TippyPopover";
-import { getHelpDocsUrl } from "metabase-lib/expressions/helper-text-strings";
 import { HelpText } from "metabase-lib/expressions/types";
+import { getHelpDocsUrl } from "./ExpressionEditorTextfield/helper-text-strings";
 
 interface ExpressionEditorHelpTextProps {
   helpText: HelpText;
