+++ frontend/src/metabase/query_builder/components/expressions/ExpressionEditorTextfield/ExpressionEditorTextfield.jsx
@@ -9,21 +9,20 @@ import AceEditor from "react-ace";
 import * as ace from "ace-builds/src-noconflict/ace";
 import ExplicitSize from "metabase/components/ExplicitSize";
 import { format } from "metabase-lib/expressions/format";
-import { suggest } from "metabase-lib/expressions/suggest";
 import { processSource } from "metabase-lib/expressions/process";
 import { diagnose } from "metabase-lib/expressions/diagnostics";
 import { tokenize } from "metabase-lib/expressions/tokenizer";
 
 import { isExpression } from "metabase-lib/expressions";
+import HelpText from "../ExpressionEditorHelpText";
+import ExpressionEditorSuggestions from "../ExpressionEditorSuggestions";
+import ExpressionMode from "../ExpressionMode";
+import { suggest } from "./suggest";
 
-import HelpText from "./ExpressionEditorHelpText";
-import ExpressionEditorSuggestions from "./ExpressionEditorSuggestions";
 import {
   EditorContainer,
   EditorEqualsSign,
 } from "./ExpressionEditorTextfield.styled";
-
-import ExpressionMode from "./ExpressionMode";
 
 ace.config.set("basePath", "/assets/ui/");
