+++ frontend/src/metabase/query_builder/components/expressions/ExpressionEditorTextfield/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./ExpressionEditorTextfield";
