+++ frontend/src/metabase/query_builder/components/expressions/ExpressionEditorTextfield/suggest.unit.spec.js
@@ -1,8 +1,10 @@
 import _ from "underscore";
 import { ORDERS, REVIEWS } from "__support__/sample_database_fixture";
-import { suggest as suggest_ } from "metabase-lib/expressions/suggest";
-
-import { aggregationOpts, expressionOpts } from "./__support__/expressions";
+import {
+  aggregationOpts,
+  expressionOpts,
+} from "../../../../../../test/metabase/lib/expressions/__support__/expressions";
+import { suggest as suggest_ } from "./suggest";
 
 // custom metadata defined in __support__/sample_database_fixture
 const SEGMENTS_ORDERS = [{ text: "[Expensive Things]", type: "segments" }];
